chore: remove conn_event and add event.rs
