+++ contracts/cosmwasm-vm/cw-ibc-core/src/ics03_connection/mod.rs
@@ -1,5 +1,4 @@
 //! ICS 03: Connection implementation for connecting a client
-pub mod conn_event;
 pub mod connection;
 pub mod event;
 use crate::context::CwIbcCoreContext;
